frontend/streamlit-application/requirements.txt
boto3~=1.35.81
streamlit~=1.41.0
streamlit-cognito-auth~=1.3.1
langfuse~=2.57.5
boto3-stubs[bedrock,bedrock-agent-runtime,s3]~=1.35.81
python-dotenv==1.0.1
python-docx==1.1.2